// apps/web/src/api/models/orderSummaryResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Autodor.API | v1
 * OpenAPI spec version: 1.0.0
 */

export interface OrderSummaryResponse {
  id: string;
  /** @nullable */
  number: string | null;
  date: string;
  /** @nullable */
  person: string | null;
  /** @nullable */
  customerNumber: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  itemsCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  totalAmount: number | string;
  isExcluded: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  excludedItemsCount: number | string;
}